Move perspective and perspective-origin properties from effects to box.
